Add CI workflow for tests and formatting
- Add CI workflow using GitHub Actions to run tests and formatting checks.
- Update `flake.nix` to include `runTests` script running from project root.
- Update `tests/Makefile` to support building both `test_main.cpp` and `test_ConnectivityManager.cpp` from project root.
- Add `test_main.cpp` for basic SystemContext tests.
- Resolve conflicts with main branch by adopting root-relative build strategy.
- Fix mock interfaces (WebServer, HTTPUpdateServer) and Arduino/mocks integration.
- Fix formatting warning in ConnectivityManager.cpp.

+++ src/ConnectivityManager.cpp
@@ -314,7 +314,7 @@ void ConnectivityManager::handleFileUpload() {
   } else if (upload.status == UPLOAD_FILE_END) {
     if (fsUploadFile) {
       fsUploadFile.close();
-      Log.printf("Upload End: %u bytes\n", upload.totalSize);
+      Log.printf("Upload End: %lu bytes\n", (unsigned long)upload.totalSize);
 
       // Hot-reload sound assets if the config file was just uploaded
       if (upload.filename.endsWith("sound_assets.json")) {
